getAllObs gets all observed properties, obsId (observed property id) now recognized properly

+++ backend/src/main/java/edu/teco/sensordatenbankmanagementsystem/models/Observation.java
@@ -59,7 +59,7 @@ public class Observation {
     Integer type;
 
     @Column(name = "\"RESULT_NUMBER\"")
-    Float resultNumber;
+    Double resultNumber;
 
     @Column(name = "\"RESULT_STRING\"")
     String resultString;
